[StockManagementSystem-4] : package create and divide.
