+++ src/main/java/jpa/jpa_shop/domain/item/Book.java
@@ -1,6 +1,6 @@
 package jpa.jpa_shop.domain.item;
 
-import jpa.jpa_shop.web.controller.dto.response.BookUpdateResponseDto;
+import jpa.jpa_shop.web.controller.dto.response.item.BookUpdateResponseDto;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
