+++ src/main/java/jpa/jpa_shop/web/controller/ItemController.java
@@ -5,12 +5,12 @@ import jpa.jpa_shop.domain.item.Book;
 import jpa.jpa_shop.domain.item.Movie;
 import jpa.jpa_shop.exception.NotSearchId;
 import jpa.jpa_shop.service.IFS.ItemServiceIFS;
-import jpa.jpa_shop.web.controller.dto.request.AlbumSaveRequestDto;
-import jpa.jpa_shop.web.controller.dto.request.BookSaveRequestDto;
-import jpa.jpa_shop.web.controller.dto.request.MovieSaveRequestDto;
-import jpa.jpa_shop.web.controller.dto.response.AlbumUpdateResponseDto;
-import jpa.jpa_shop.web.controller.dto.response.BookUpdateResponseDto;
-import jpa.jpa_shop.web.controller.dto.response.MovieUpdateResponseDto;
+import jpa.jpa_shop.web.controller.dto.request.item.AlbumSaveRequestDto;
+import jpa.jpa_shop.web.controller.dto.request.item.BookSaveRequestDto;
+import jpa.jpa_shop.web.controller.dto.request.item.MovieSaveRequestDto;
+import jpa.jpa_shop.web.controller.dto.response.item.AlbumUpdateResponseDto;
+import jpa.jpa_shop.web.controller.dto.response.item.BookUpdateResponseDto;
+import jpa.jpa_shop.web.controller.dto.response.item.MovieUpdateResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
