+++ src/main/java/jpa/jpa_shop/web/controller/dto/request/member/MemberSaveRequestDto.java
@@ -1,4 +1,4 @@
-package jpa.jpa_shop.web.controller.dto.request;
+package jpa.jpa_shop.web.controller.dto.request.member;
 
 import jpa.jpa_shop.domain.member.Address;
 import jpa.jpa_shop.domain.member.Member;
