+++ src/test/java/jpa/jpa_shop/service/OrderServiceTest.java
@@ -7,7 +7,7 @@ import jpa.jpa_shop.domain.member.Member;
 import jpa.jpa_shop.domain.orders.Order;
 import jpa.jpa_shop.domain.orders.OrderStatus;
 import jpa.jpa_shop.domain.orders.Repository.OrderRepository;
-import jpa.jpa_shop.web.controller.dto.request.OrderSearchRequestDto;
+import jpa.jpa_shop.web.controller.dto.request.order.OrderSearchRequestDto;
 import jpa.jpa_shop.exception.NotEnoughStockException;
 import jpa.jpa_shop.service.IFS.OrderServiceIFS;
 import org.assertj.core.api.Assertions;
